+++ Source/DafnyCore/Compilers/Compiler-cpp.cs
@@ -9,21 +9,17 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
-using System.IO;
 using System.Diagnostics.Contracts;
 using System.Collections.ObjectModel;
-using System.Diagnostics;
 using JetBrains.Annotations;
 
 namespace Microsoft.Dafny.Compilers {
-  public class CppCompiler : SinglePassCompiler {
-    public override void OnPreCompile(ErrorReporter reporter, ReadOnlyCollection<string> otherFileNames) {
-      base.OnPreCompile(reporter, otherFileNames);
-      datatypeDecls = new List<DatatypeDecl>();
-      classDefaults = new List<string>();
-    }
+  class CppCompiler : SinglePassCompiler {
 
-    public override IReadOnlySet<string> SupportedExtensions => new HashSet<string> { ".h" };
+    private ReadOnlyCollection<string> headers;
+    public CppCompiler(ErrorReporter reporter, ReadOnlyCollection<string> headers) : base(reporter) {
+      this.headers = headers;
+    }
 
     public override IReadOnlySet<Feature> UnsupportedFeatures => new HashSet<Feature> {
       Feature.UnboundedIntegers,
@@ -59,16 +55,9 @@ namespace Microsoft.Dafny.Compilers {
       Feature.UnicodeChars,
       Feature.ConvertingValuesToStrings
     };
-
-    public override string TargetLanguage => "C++";
-    public override string TargetExtension => "cpp";
-
-    public override bool SupportsInMemoryCompilation => false;
-    public override bool TextualTargetIsExecutable => false;
-
-    private ReadOnlyCollection<string> headers => OtherFileNames;
-    private List<DatatypeDecl> datatypeDecls;
-    private List<string> classDefaults;
+    
+    private List<DatatypeDecl> datatypeDecls = new();
+    private List<string> classDefaults = new();
 
     /*
      * Unlike other Dafny and Dafny's other backends, C++ cares about
@@ -129,7 +118,6 @@ namespace Microsoft.Dafny.Compilers {
       var rt = wr.NewFile("DafnyRuntime.h");
       ReadRuntimeSystem(program, "DafnyRuntime.h", rt);
     }
-
     protected override void EmitFooter(Program program, ConcreteSyntaxTree wr) {
       // Define default values for each datatype
       foreach (var dt in this.datatypeDecls) {
@@ -160,7 +148,7 @@ namespace Microsoft.Dafny.Compilers {
     }
 
     protected override ConcreteSyntaxTree CreateStaticMain(IClassWriter cw, string argsParameterName) {
-      var wr = (cw as CppCompiler.ClassWriter).MethodWriter;
+      var wr = (cw as ClassWriter).MethodWriter;
       return wr.NewBlock($"int main(DafnySequence<DafnySequence<char>> {argsParameterName})");
     }
 
@@ -621,7 +609,7 @@ namespace Microsoft.Dafny.Compilers {
         throw new UnsupportedFeatureException(nt.tok, Feature.NonNativeNewtypes, String.Format("non-native newtype {0}", nt));
       }
       var className = "class_" + IdName(nt);
-      var cw = CreateClass(nt.EnclosingModuleDefinition.CompileName, className, nt, wr) as CppCompiler.ClassWriter;
+      var cw = CreateClass(nt.EnclosingModuleDefinition.CompileName, className, nt, wr) as ClassWriter;
       var w = cw.MethodDeclWriter;
       if (nt.WitnessKind == SubsetTypeDecl.WKind.Compiled) {
         var witness = new ConcreteSyntaxTree(w.RelativeIndentLevel);
@@ -661,7 +649,7 @@ namespace Microsoft.Dafny.Compilers {
       this.modDeclWr.WriteLine("{0} using {1} = {2};", templateDecl, IdName(sst), TypeName(sst.Var.Type, wr, sst.tok));
 
       var className = "class_" + IdName(sst);
-      var cw = CreateClass(sst.EnclosingModuleDefinition.CompileName, className, sst, wr) as CppCompiler.ClassWriter;
+      var cw = CreateClass(sst.EnclosingModuleDefinition.CompileName, className, sst, wr) as ClassWriter;
       var w = cw.MethodDeclWriter;
 
       if (sst.WitnessKind == SubsetTypeDecl.WKind.Compiled) {
@@ -1563,6 +1551,7 @@ namespace Microsoft.Dafny.Compilers {
     protected override string IdProtect(string name) {
       return PublicIdProtect(name);
     }
+    
     public override string PublicIdProtect(string name) {
       Contract.Requires(name != null);
       switch (name) {
@@ -2473,39 +2462,5 @@ namespace Microsoft.Dafny.Compilers {
 
 
     // ----- Target compilation and execution -------------------------------------------------------------
-    private string ComputeExeName(string targetFilename) {
-      return Path.ChangeExtension(Path.GetFullPath(targetFilename), "exe");
-    }
-
-    public override bool CompileTargetProgram(string dafnyProgramName, string targetProgramText, string/*?*/ callToMain, string/*?*/ targetFilename, ReadOnlyCollection<string> otherFileNames,
-      bool runAfterCompile, TextWriter outputWriter, out object compilationResult) {
-      var assemblyLocation = System.Reflection.Assembly.GetExecutingAssembly().Location;
-      Contract.Assert(assemblyLocation != null);
-      var codebase = System.IO.Path.GetDirectoryName(assemblyLocation);
-      Contract.Assert(codebase != null);
-      compilationResult = null;
-      var psi = PrepareProcessStartInfo("g++", new List<string> {
-        "-Wall",
-        "-Wextra",
-        "-Wpedantic",
-        "-Wno-unused-variable",
-        "-Wno-deprecated-copy",
-        "-Wno-unused-label",
-        "-Wno-unused-but-set-variable",
-        "-Wno-unknown-warning-option",
-        "-g",
-        "-std=c++17",
-        "-I", codebase,
-        "-o", ComputeExeName(targetFilename),
-        targetFilename
-      });
-      return 0 == RunProcess(psi, outputWriter, "Error while compiling C++ files.");
-    }
-
-    public override bool RunTargetProgram(string dafnyProgramName, string targetProgramText, string/*?*/ callToMain, string targetFilename, ReadOnlyCollection<string> otherFileNames,
-      object compilationResult, TextWriter outputWriter) {
-      var psi = PrepareProcessStartInfo(ComputeExeName(targetFilename), DafnyOptions.O.MainArgs);
-      return 0 == RunProcess(psi, outputWriter);
-    }
   }
 }
