+++ Source/DafnyCore/Compilers/SinglePassCompiler.cs
@@ -10,9 +10,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using System.IO;
-using System.Diagnostics;
 using System.Diagnostics.Contracts;
-using System.Collections.ObjectModel;
 using JetBrains.Annotations;
 using Microsoft.BaseTypes;
 
@@ -23,24 +21,46 @@ namespace Microsoft.Dafny.Compilers {
 
     public override Plugins.Compiler[] GetCompilers() {
       return new Plugins.Compiler[] {
-        new Compilers.CsharpCompiler(),
+        new Compilers.CsharpBackend(),
         new Compilers.JavaScriptCompiler(),
         new Compilers.GoCompiler(),
         new Compilers.JavaCompiler(),
         new Compilers.PythonCompiler(),
-        new Compilers.CppCompiler()
+        new Compilers.CppCompilerBackend()
       };
     }
   }
 
-  public abstract class SinglePassCompiler : Plugins.Compiler {
+  public abstract class SinglePassCompiler {
+    
+    /// <summary>
+    /// Emits a call to <c>mainMethod</c> as the program's entry point, if such an explicit call is
+    /// required in the target language.
+    /// </summary>
+    public abstract void EmitCallToMain(Method mainMethod, string baseName, ConcreteSyntaxTree callToMainTree);
+    
+    /// <summary>
+    /// Change <c>name</c> into a valid identifier in the target language.
+    /// </summary>
+    public abstract string PublicIdProtect(string name);
+    
     public static Plugin Plugin =
       new ConfiguredPlugin(InternalCompilersPluginConfiguration.Singleton);
 
+    public abstract IReadOnlySet<Feature> UnsupportedFeatures { get; }
+
+    /// <summary>
+    /// Whether or not the compiler turns
+    ///     datatype Record = R(oneThing: X)
+    /// into just X, including the case where "Record" is a tuple type with 1 non-ghost component.
+    /// </summary>
+    public virtual bool SupportsDatatypeWrapperErasure => true;
     public static string DefaultNameMain = "Main";
 
     protected virtual string ModuleSeparator { get => "."; }
     protected virtual string ClassAccessor { get => "."; }
+    
+    protected ErrorReporter? Reporter;
 
     Stack<ConcreteSyntaxTree> copyInstrWriters = new Stack<ConcreteSyntaxTree>(); // a buffer that stores copy instructions generated by letExpr that uses out param.
     protected TopLevelDeclWithMembers thisContext;  // non-null when type members are being translated
@@ -65,45 +85,10 @@ namespace Microsoft.Dafny.Compilers {
 
     public CoverageInstrumenter Coverage;
 
-    public ProcessStartInfo PrepareProcessStartInfo(string programName, IEnumerable<string> args = null) {
-      var psi = new ProcessStartInfo(programName) {
-        UseShellExecute = false,
-        CreateNoWindow = false, // https://github.com/dotnet/runtime/issues/68259
-      };
-      foreach (var arg in args ?? Enumerable.Empty<string>()) {
-        psi.ArgumentList.Add(arg);
-      }
-      return psi;
+    protected SinglePassCompiler(ErrorReporter reporter) {
+      Reporter = reporter;
+      Coverage = new CoverageInstrumenter(this);
     }
-
-    public int WaitForExit(Process process, TextWriter outputWriter, string errorMessage = null) {
-      process.WaitForExit();
-      if (process.ExitCode != 0 && errorMessage != null) {
-        outputWriter.WriteLine("{0} Process exited with exit code {1}", errorMessage, process.ExitCode);
-      }
-      return process.ExitCode;
-    }
-
-    public Process StartProcess(ProcessStartInfo psi, TextWriter outputWriter) {
-      string additionalInfo = "";
-
-      try {
-        if (Process.Start(psi) is { } process) {
-          return process;
-        }
-      } catch (System.ComponentModel.Win32Exception e) {
-        additionalInfo = $": {e.Message}";
-      }
-
-      outputWriter.WriteLine($"Error: Unable to start {psi.FileName}{additionalInfo}");
-      return null;
-    }
-
-    public int RunProcess(ProcessStartInfo psi, TextWriter outputWriter, string errorMessage = null) {
-      return StartProcess(psi, outputWriter) is { } process ?
-         WaitForExit(process, outputWriter, errorMessage) : -1;
-    }
-
 
     protected static void ReportError(ErrorReporter reporter, IToken tok, string msg, ConcreteSyntaxTree/*?*/ wr, params object[] args) {
       Contract.Requires(msg != null);
@@ -119,7 +104,7 @@ namespace Microsoft.Dafny.Compilers {
 
     protected void UnsupportedFeatureError(IToken tok, Feature feature, string message = null, ConcreteSyntaxTree wr = null, params object[] args) {
       if (!UnsupportedFeatures.Contains(feature)) {
-        throw new Exception($"'{feature}' is not an element of the {TargetId} compiler's UnsupportedFeatures set");
+        throw new Exception($"'{feature}' is not an element of the {GetType().Name} compiler's UnsupportedFeatures set");
       }
 
       message ??= UnsupportedFeatureException.MessagePrefix + FeatureDescriptionAttribute.GetDescription(feature).Description;
@@ -136,17 +121,6 @@ namespace Microsoft.Dafny.Compilers {
     protected virtual void EmitHeader(Program program, ConcreteSyntaxTree wr) { }
     protected virtual void EmitFooter(Program program, ConcreteSyntaxTree wr) { }
     protected virtual void EmitBuiltInDecls(BuiltIns builtIns, ConcreteSyntaxTree wr) { }
-
-
-    public override void OnPreCompile(ErrorReporter reporter, ReadOnlyCollection<string> otherFileNames) {
-      base.OnPreCompile(reporter, otherFileNames);
-      Coverage = new CoverageInstrumenter(this);
-    }
-
-    public override void OnPostCompile() {
-      base.OnPostCompile();
-      Coverage.WriteLegendFile();
-    }
 
     /// <summary>
     /// Creates a static Main method. The caller will fill the body of this static Main with a
@@ -1351,7 +1325,7 @@ namespace Microsoft.Dafny.Compilers {
       modules = program.CompileModules;
     }
 
-    public override void Compile(Program program, ConcreteSyntaxTree wrx) {
+    public void Compile(Program program, ConcreteSyntaxTree wrx) {
       Contract.Requires(program != null);
 
       EmitHeader(program, wrx);
@@ -5712,29 +5686,6 @@ namespace Microsoft.Dafny.Compilers {
 
     protected abstract void EmitHaltRecoveryStmt(Statement body, string haltMessageVarName, Statement recoveryBody, ConcreteSyntaxTree wr);
 
-    public override bool CompileTargetProgram(string dafnyProgramName, string targetProgramText, string/*?*/ callToMain, string/*?*/ targetFilename, ReadOnlyCollection<string> otherFileNames,
-      bool runAfterCompile, TextWriter outputWriter, out object compilationResult) {
-      Contract.Requires(dafnyProgramName != null);
-      Contract.Requires(targetProgramText != null);
-      Contract.Requires(otherFileNames != null);
-      Contract.Requires(otherFileNames.Count == 0 || targetFilename != null);
-      Contract.Requires(this.SupportsInMemoryCompilation || targetFilename != null);
-      Contract.Requires(!runAfterCompile || callToMain != null);
-      Contract.Requires(outputWriter != null);
-
-      compilationResult = null;
-      return true;
-    }
-
-    public override bool RunTargetProgram(string dafnyProgramName, string targetProgramText, string/*?*/ callToMain, string/*?*/ targetFilename, ReadOnlyCollection<string> otherFileNames,
-      object compilationResult, TextWriter outputWriter) {
-      Contract.Requires(dafnyProgramName != null);
-      Contract.Requires(targetProgramText != null);
-      Contract.Requires(otherFileNames != null);
-      Contract.Requires(otherFileNames.Count == 0 || targetFilename != null);
-      Contract.Requires(outputWriter != null);
-      return true;
-    }
   }
 
   public class CoverageInstrumenter {
