+++ Source/DafnyCore/Plugins/Compiler.cs
@@ -76,13 +76,6 @@ public abstract class Compiler {
   public abstract bool SupportsInMemoryCompilation { get; }
 
   /// <summary>
-  /// Whether or not the compiler turns
-  ///     datatype Record = R(oneThing: X)
-  /// into just X, including the case where "Record" is a tuple type with 1 non-ghost component.
-  /// </summary>
-  public virtual bool SupportsDatatypeWrapperErasure => true;
-
-  /// <summary>
   /// Dafny features this compiler is known to not support.
   /// </summary>
   public virtual IReadOnlySet<Feature> UnsupportedFeatures => new HashSet<Feature>();
